Implement vsync

+++ src/renderer/vulkan/shader/shader.c
@@ -240,7 +240,8 @@ void shader_use(VulkanShader *shader) {
       VK_PIPELINE_BIND_POINT_GRAPHICS, &shader->pipeline);
 }
 
-void vulkan_shader_update_global_state(VulkanShader *shader, float delta_time) {
+void vulkan_shader_update_global_state(
+  VulkanShader *shader, double delta_time) {
     unsigned int image_index = vulkan_context.image_index;
     VkCommandBuffer command_buffer =
       vulkan_context.graphics_command_buffers[image_index].handle;
