+++ src/renderer/vulkan/shader/shader.h
@@ -8,7 +8,7 @@ void shader_destroy(VulkanShader *shader);
 
 void shader_use(VulkanShader *shader);
 
-void vulkan_shader_update_global_state(VulkanShader *shader, float delta_time);
+void vulkan_shader_update_global_state(VulkanShader *shader, double delta_time);
 
 void vulkan_shader_update_object(VulkanShader *shader, GeometryRenderData data);
